requirements.txt
matplotlib==3.7.4
numpy==1.21.5
scipy==1.7.3
pandas==1.4.1
astropy==5.0.4
tqdm==4.62.3
git+https://github.com/dsa110/dsa110-antpos.git
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
casatools==6.4.4.31
pytest==7.0.1